Add reschedule command
The reschedule command takes a name and a datetime. The job is then
rescheduled to run next at that datetime. The command internally allows
for changing the period as well, but this is not exposed in the client
command (yet).

The semantics of rescheduling to a time which is now or in the past is
the same as running execute.