Add reschedule command (#8)
* Add reschedule command

The reschedule command takes a name and a datetime. The job is then
rescheduled to run next at that datetime. The command internally allows
for changing the period as well,.

The semantics of rescheduling to a time which is now or in the past is
the same as running execute.